add simple solver flow for further development

+++ include/Point.h
@@ -7,7 +7,7 @@ extern "C" {
 
 typedef struct Point
 {
-    int row, column;
+    size_t row, column;
 } Point;
 
 #ifdef __cplusplus
